feat: Add new components for Songs, AboutUs, Categories, Overview, Home, and NotFound
This commit adds new components for Songs, AboutUs, Categories, Overview, Home, and NotFound. These components are essential for building the different pages of the application. Each component is a functional component that returns a div element with the respective page content.

The components are added to the client/src/components directory. They are imported and exported correctly to be used in other parts of the application.

This commit also includes the necessary changes to the package.json file to add new dependencies. The react-router and react-router-dom packages are added to enable routing functionality in the application. The react-spinners package is added to provide loading spinners for better user experience.

Lastly, this commit includes the addition of new image files for the application. The Vector.svg and Triangle.svg files are added to the client/src/assets/images directory. The Mask group.png and Mask group-1.png files are added to the client/public directory.

These changes are made to enhance the functionality and appearance of the application.